requirements.txt
# Lung Cancer Detection API — CPU-only dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Core
numpy==1.24.3
scipy==1.11.1

# ML
scikit-learn==1.3.0
xgboost==2.0.0
imbalanced-learn==0.11.0

# Deep Learning (CPU-only — much smaller than CUDA builds)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
torchvision==0.16.0+cpu

# Quantum
pennylane==0.32.0
pennylane-lightning==0.32.0

# Image processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# Config
pyyaml==6.0.1
joblib==1.3.2
